Hardcode LC_CTYPE for lowerstr_with_len
This is what the production database's Ctype was previously set to. It
affects the dictionary normalization, because `lowerstr_with_len` doesn't
handle wide chars when the "C" locale is used (i.e. what the production
database is currently using).

So, the current behavior is that `musicbrainz_dunaccentdict_lexize`
doesn't lowercase any wide chars. While it would help in some cases
to reverse the order in which `lowerstr_with_len` and `unaccent_string`
are called (e.g. "Æ" could unaccent to AE, which `lowerstr_with_len`
could then handle even in the "C" locale), that doesn't work for wide
chars which can be lowercased but not unaccented (e.g. "П"). That
change is not as simple as reversing the calls, anyway.

Hardcoding the locale isn't exactly ideal, but it restores the previous
dictionary normalization behavior, and we can't know the actual context
(locale) of the strings being passed in anyway.